refactor: restore old stake page until its finished
